changed dispatch function to dispatch on the prediction network type
